+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   in specified databases.
 - Option `--coverage-fail-under` to `odoo-helper test` command
 - Option `--except-filter` to `odoo-helper addons list` command
+- Command `odoo-helper ci ensure-changelog` command
 
 ### Changed